refactor(palette): extract useCommandPalette hook + virtualize listbox
P2 of the architecture refactor. Move all derived state, effects, and a
flattened heading+option row model out of CommandPalette.tsx into
hooks/useCommandPalette.ts; the component is now purely presentational
and virtualizes the listbox via @tanstack/react-virtual (same pattern as
NavigatorPanel). The custom WCAG-tuned fuzzy scorer is retained — no new
dependency, no ranking/highlight behavior change.

Keyboard nav (flatItems index space) and the virtualized row list (rows
index space, incl. headings) are bridged by optionIndexToRowIndex so the
selected option scrolls into view. ARIA combobox/listbox/option semantics
and aria-activedescendant preserved.

Tests: existing 4 smoke tests retained; added a deterministic
renderHook test asserting the flattened row model + option-index map.
Verified: typecheck, lint (0 warnings), 5 CommandPalette tests pass.
